+++ CHANGELOG.md
@@ -1,1 +1,19 @@
 # Change Log
+
+## v0.1.1
+
+### Features
+
++ rss feed support `https`
++ proxy support `https`,`socks`
++ serve support static files
+
+### Bugfixes
+
++ feed `https` cache error
+
+### Others
+
++ migrate actix to wrap
++ refactor info message
++ skip failed rss feed